+++ lib/rsyntaxtree/string_parser.rb
@@ -41,7 +41,10 @@ module RSyntaxTree
       end
 
       @data = string # Store it for later...
-      fontset[:normal] = fontset[:cjk] if @data.contains_cjk?
+      if @data.contains_cjk?
+        fontset[:normal] = fontset[:cjk]
+        fontset[:family] = fontset[:family_cjk]
+      end
       @elist = ElementList.new # Initialize internal element list
       @pos = 0 # Position in the sentence
       @id = 1 # ID for the next element